+++ test_result.go
@@ -38,6 +38,7 @@ func (m buildTestResultModule) Register(r gin.IRouter) {
 // createBuildTestResultHandler godoc
 // @id createBuildTestResult
 // @summary Post test result data
+// @description Added in v5.0.0.
 // @tags test-result
 // @accept multipart/form-data
 // @param buildId path uint true "Build ID" minimum(0)
@@ -125,6 +126,7 @@ func (m buildTestResultModule) createBuildTestResultHandler(c *gin.Context) {
 // getBuildAllTestResultDetailListHandler godoc
 // @id getBuildAllTestResultDetailList
 // @summary Get all test result details for specified build
+// @description Added in v5.0.0.
 // @tags test-result
 // @param buildId path uint true "Build ID" minimum(0)
 // @success 200 {object} response.PaginatedTestResultDetails
@@ -160,6 +162,7 @@ func (m buildTestResultModule) getBuildAllTestResultDetailListHandler(c *gin.Con
 // getBuildAllTestResultSummaryListHandler godoc
 // @id getBuildAllTestResultSummaryList
 // @summary Get all test result summaries for specified build
+// @description Added in v5.0.0.
 // @tags test-result
 // @param buildId path uint true "Build ID" minimum(0)
 // @success 200 {object} response.PaginatedTestResultSummaries
@@ -199,6 +202,7 @@ func (m buildTestResultModule) getBuildAllTestResultSummaryListHandler(c *gin.Co
 // getBuildTestResultSummaryHandler godoc
 // @id getBuildTestResultSummary
 // @summary Get test result summary for specified test
+// @description Added in v5.0.0.
 // @tags test-result
 // @param buildId path uint true "Build ID" minimum(0)
 // @param artifactId path uint true "Artifact ID" minimum(0)
@@ -237,6 +241,7 @@ func (m buildTestResultModule) getBuildTestResultSummaryHandler(c *gin.Context) 
 // getBuildTestResultDetailListHandler godoc
 // @id getBuildTestResultDetailList
 // @summary Get all test result details for specified test
+// @description Added in v5.0.0.
 // @tags test-result
 // @param buildId path uint true "Build ID" minimum(0)
 // @param artifactId path uint true "Artifact ID" minimum(0)
@@ -278,6 +283,7 @@ func (m buildTestResultModule) getBuildTestResultDetailListHandler(c *gin.Contex
 // getBuildAllTestResultListSummaryHandler godoc
 // @id getBuildAllTestResultListSummary
 // @summary Get test result list summary of all tests for specified build
+// @description Added in v5.0.0.
 // @tags test-result
 // @param buildId path uint true "Build ID" minimum(0)
 // @success 200 {object} response.TestResultListSummary